refactor: useLoadProduct로 fetching하는 로직 훅으로 분리
