+++ src/pages/ProductListPage/index.tsx
@@ -1,80 +1,26 @@
-import { fetchProduct } from '@apis/index';
-import { CartItem, Filtering, Product } from '@appTypes/index';
+import { CartItem, Filtering } from '@appTypes/index';
 import { Dropdown, IntersectionObserverArea } from '@components/index';
 import { CATEGORY_OPTIONS, PRICE_SORT_OPTIONS } from '@constants/index';
 import { useEffect, useRef, useState } from 'react';
 
 import ProductList from './ProductList';
 import style from './style.module.css';
+import useLoadProducts from '@src/hooks/useLoadProducts';
 
 interface ProductListPageProps {
   cartItems: CartItem[];
 }
 
 function ProductListPage({ cartItems }: ProductListPageProps) {
-  const [products, setProducts] = useState<Product[]>([]);
-  const [page, setPage] = useState(0);
-  const [isLast, setIsLast] = useState(false);
   const [filtering, setFiltering] = useState<Filtering>({ category: '', sort: 'price,asc' });
+  const { products, loading, error, loadNextPage } = useLoadProducts(filtering);
   const targetRef = useRef<HTMLDivElement | null>(null);
-
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState('');
-
-  const updateState = ({
-    isLast,
-    newProducts,
-    newPage,
-  }: {
-    isLast: boolean;
-    newProducts: Product[];
-    newPage: number;
-  }) => {
-    setIsLast(isLast);
-    setProducts(newProducts);
-    setPage(newPage);
-  };
-
-  const callback = async () => {
-    try {
-      if (isLast || !products.length) return;
-      setLoading(true);
-      const newPage = page + 1;
-      const result = await fetchProduct({ filtering, page: newPage });
-      if (!result) return;
-      updateState({ newPage, ...result, newProducts: [...products, ...result.products], isLast: result.isLast });
-      setError('');
-    } catch (error) {
-      const errorMessage = (error as Error).message;
-      setError(errorMessage);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleFilteringProducts = async () => {
-    try {
-      setProducts([]);
-      setLoading(true);
-      const result = await fetchProduct({ filtering });
-      updateState({ isLast: result.isLast, newProducts: result.products, newPage: 0 });
-    } catch (error) {
-      const errorMessage = (error as Error).message;
-      setError(errorMessage);
-    } finally {
-      setLoading(false);
-    }
-  };
 
   const handleChangeOption = (event: React.ChangeEvent<HTMLSelectElement>) => {
     const { name, value } = event.target;
 
     setFiltering((prev) => ({ ...prev, [name]: value }));
   };
-
-  useEffect(() => {
-    handleFilteringProducts();
-  }, [filtering]);
 
   useEffect(() => {
     if (error) {
@@ -89,7 +35,7 @@ function ProductListPage({ cartItems }: ProductListPageProps) {
         <Dropdown label="카테고리" name="category" options={CATEGORY_OPTIONS} onChange={handleChangeOption} />
         <Dropdown label="가격순" name="sort" options={PRICE_SORT_OPTIONS} onChange={handleChangeOption} />
       </div>
-      <IntersectionObserverArea callback={callback} targetRef={targetRef}>
+      <IntersectionObserverArea callback={loadNextPage} targetRef={targetRef}>
         <ProductList products={products} targetRef={targetRef} loading={loading} cartItems={cartItems} />
       </IntersectionObserverArea>
     </div>
